+++ src/webview/parser/statements/index.ts
@@ -8,8 +8,9 @@ export {
     tryProcessCreateStatement,
     type StatementPresentation
 } from './ddl';
+export { tryParseCompatibleDeleteStatement } from './delete';
 export { tryParseBulkDataStatement } from './bulk';
-export { tryProcessDmlStatements, type ProcessDmlStatementsArgs } from './dml';
+export { resolveDeleteTargetTableNames, tryProcessDmlStatements, type ProcessDmlStatementsArgs } from './dml';
 export { tryParseCompatibleMergeStatement } from './merge';
 export {
     processSelectStatement,